requirements.txt
asgiref==3.8.1
boto3==1.35.76
botocore==1.35.76
Django==5.1.4
django-cors-headers==4.6.0
django-rest-authtoken==2.1.4
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.2
urllib3==2.2.3
whitenoise==6.8.2
